crud on server

+++ js/main.js
@@ -1,8 +1,11 @@
 $(document).ready(function () {
+    var starsPlugin = $('[name="review"]').stars();
+
     var drawTable = function (store) {
         store.getAll().then(function (data) {
             $('tbody').empty();
-            $.each(data, function(){
+            $.each(data.list, function (index) {
+                data.list[index].stars = starsPlugin.activeStars(data.list[index].stars);
                 var sc = $('#tpl').html();
                 var s = tmpl(sc, this);
                 $("tbody").append(s);
@@ -13,31 +16,80 @@ $(document).ready(function () {
 
     drawTable(store);
 
+    var resetForm = function() {
+        $('.form-control').val('');
+        $('[name="visited"]').attr('checked', false);
+        $('[name="iou"]').val('0');
+        $('[name="review"]').val('0').change();
+    }
+
     $('#my-form').submit(function () {
-        store.add(getVal()).then(function (data) {
-            drawTable(store);
-            console.log(data);
-        });
+        var iou = $('[name="iou"]').val();
+        if (iou == '0') {
+            store.add(getVal()).then(
+                function (data) {
+                    drawTable(store);
+                    resetForm();
+                },
+                function (data) {
+                    alert(data.error);
+                }
+             );
+        } else {
+            index = iou;
+            store.update(index, getVal()).then(
+                function (data) {
+                    drawTable(store);
+                    resetForm();
+                },
+                function (data) {
+                    alert(data.error);
+                }
+            );
+        };
         return false;
     });
 
     var getVal = function() {
         return {
-            city: $('#city').val(),
-            stars: 3,
-            visited: $('#cb:checkbox:checked').length > 0
+            name: $('#city').val(),
+            stars: parseInt($('.stars').val()),
+            visited: $('#cb:checkbox:checked').length > 0 ? 1 : 0
         }
     };
 
-    var attachEvents = function() {
-        $('.remove-btn').click(function (){
-            var rowId = $(this).closest('tr').data('id');
-            store.delete(rowId).then(function() {
-               drawTable(store);
-           });
+    var removeRow = function() {
+        $('.remove-btn').confirm({
+            message: 'Are you sure?',
+            onConfirm: function() {
+                console.log(this, 'yes!');
+                var rowId = $(this).closest('tr').data('id');
+                store.delete(rowId).then(function() {
+                    drawTable(store);
+                });
+            }
         });
-    }
+    };
+
+    var editRow = function() {
+        $('.edit-btn').click(function() {
+            var rowId = $(this).closest('tr').data('id');
+            store.get(rowId).then(function(data) {
+                        $('#city').val(data.name);
+                        $('#cb').prop('checked', data.visited);
+                        $('.stars').val(data.stars).change();
+                        $('[name="iou"]').val(data.id);
+                });
+            });
+        };
+
+    var attachEvents = function() {
+        removeRow();
+        editRow();
+    };
+
 });
+
